// src/types/school.ts
export enum SchoolType {
  PLAY_SCHOOL = "PLAY_SCHOOL",
  KINDERGARTEN = "KINDERGARTEN",
  PRIMARY = "PRIMARY",
  SECONDARY = "SECONDARY",
  HIGHER_SECONDARY = "HIGHER_SECONDARY",
  COLLEGE = "COLLEGE",
  UNIVERSITY = "UNIVERSITY",
  INSTITUTE = "INSTITUTE",
  OTHER = "OTHER",
}

export enum SchoolStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  ARCHIVED = "ARCHIVED",
}